+++ cairo1-run/src/cairo_run.rs
@@ -30,7 +30,10 @@ use cairo_vm::{
     hint_processor::cairo_1_hint_processor::hint_processor::Cairo1HintProcessor,
     math_utils::signed_felt,
     serde::deserialize_program::{ApTracking, FlowTrackingData, HintParams, ReferenceManager},
-    types::{builtin_name::BuiltinName, program::Program, relocatable::MaybeRelocatable},
+    types::{
+        builtin_name::BuiltinName, layout_name::LayoutName, program::Program,
+        relocatable::MaybeRelocatable,
+    },
     vm::{
         errors::{runner_errors::RunnerError, vm_errors::VirtualMachineError},
         runners::cairo_runner::{CairoRunner, RunResources, RunnerMode},
@@ -51,7 +54,7 @@ pub struct Cairo1RunConfig<'a> {
     pub serialize_output: bool,
     pub trace_enabled: bool,
     pub relocate_mem: bool,
-    pub layout: &'a str,
+    pub layout: LayoutName,
     pub proof_mode: bool,
     // Should be true if either air_public_input or cairo_pie_output are needed
     // Sets builtins stop_ptr by calling `final_stack` on each builtin
@@ -67,7 +70,7 @@ impl Default for Cairo1RunConfig<'_> {
             serialize_output: false,
             trace_enabled: false,
             relocate_mem: false,
-            layout: "plain",
+            layout: LayoutName::plain,
             proof_mode: false,
             finalize_builtins: false,
             append_return_values: false,
@@ -1159,7 +1162,7 @@ mod tests {
         // Set proof_mode
         let cairo_run_config = Cairo1RunConfig {
             proof_mode,
-            layout: "all_cairo",
+            layout: LayoutName::all_cairo,
             append_return_values: !proof_mode, // This is so we can test appending return values when not running in proof_mode
             finalize_builtins: true,
             ..Default::default()