+++ vm/src/vm/runners/cairo_runner.rs
@@ -7,14 +7,14 @@ use crate::{
         ops::{Add, AddAssign, Mul, MulAssign, Sub, SubAssign},
         prelude::*,
     },
-    types::{builtin_name::BuiltinName, layout::MEMORY_UNITS_PER_STEP},
+    types::{builtin_name::BuiltinName, layout::MEMORY_UNITS_PER_STEP, layout_name::LayoutName},
     vm::{
         runners::builtin_runner::SegmentArenaBuiltinRunner,
         trace::trace_entry::{relocate_trace_register, RelocatedTraceEntry},
     },
+    Felt252,
 };
 
-use crate::Felt252;
 use crate::{
     hint_processor::hint_processor_definition::{HintProcessor, HintReference},
     math_utils::safe_div_usize,
@@ -171,25 +171,20 @@ pub enum RunnerMode {
 impl CairoRunner {
     pub fn new_v2(
         program: &Program,
-        layout: &str,
+        layout: LayoutName,
         mode: RunnerMode,
     ) -> Result<CairoRunner, RunnerError> {
         let cairo_layout = match layout {
-            "plain" => CairoLayout::plain_instance(),
-            "small" => CairoLayout::small_instance(),
-            "dex" => CairoLayout::dex_instance(),
-            "recursive" => CairoLayout::recursive_instance(),
-            "starknet" => CairoLayout::starknet_instance(),
-            "starknet_with_keccak" => CairoLayout::starknet_with_keccak_instance(),
-            "recursive_large_output" => CairoLayout::recursive_large_output_instance(),
-            "all_cairo" => CairoLayout::all_cairo_instance(),
-            "all_solidity" => CairoLayout::all_solidity_instance(),
-            "dynamic" => CairoLayout::dynamic_instance(),
-            name => {
-                return Err(RunnerError::InvalidLayoutName(
-                    name.to_string().into_boxed_str(),
-                ))
-            }
+            LayoutName::plain => CairoLayout::plain_instance(),
+            LayoutName::small => CairoLayout::small_instance(),
+            LayoutName::dex => CairoLayout::dex_instance(),
+            LayoutName::recursive => CairoLayout::recursive_instance(),
+            LayoutName::starknet => CairoLayout::starknet_instance(),
+            LayoutName::starknet_with_keccak => CairoLayout::starknet_with_keccak_instance(),
+            LayoutName::recursive_large_output => CairoLayout::recursive_large_output_instance(),
+            LayoutName::all_cairo => CairoLayout::all_cairo_instance(),
+            LayoutName::all_solidity => CairoLayout::all_solidity_instance(),
+            LayoutName::dynamic => CairoLayout::dynamic_instance(),
         };
         Ok(CairoRunner {
             program: program.clone(),
@@ -218,7 +213,7 @@ impl CairoRunner {
 
     pub fn new(
         program: &Program,
-        layout: &str,
+        layout: LayoutName,
         proof_mode: bool,
     ) -> Result<CairoRunner, RunnerError> {
         if proof_mode {
@@ -357,7 +352,7 @@ impl CairoRunner {
         if !program_builtins.is_empty() && !allow_missing_builtins {
             return Err(RunnerError::NoBuiltinForInstance(Box::new((
                 program_builtins.iter().map(|n| **n).collect(),
-                self.layout.name.clone(),
+                self.layout.name,
             ))));
         }
 
@@ -1445,15 +1440,14 @@ impl CairoRunner {
         &self,
         vm: &VirtualMachine,
     ) -> Result<PublicInput, PublicInputError> {
-        let layout_name = self.get_layout().name.as_str();
-        let dyn_layout = match layout_name {
-            "dynamic" => Some(self.get_layout()),
+        let dyn_layout = match self.layout.name {
+            LayoutName::dynamic => Some(self.get_layout()),
             _ => None,
         };
 
         PublicInput::new(
             &self.relocated_memory,
-            layout_name,
+            self.layout.name.to_str(),
             dyn_layout,
             &vm.get_public_memory_addresses()?,
             self.get_memory_segment_addresses(vm)?,
@@ -1664,7 +1658,7 @@ mod tests {
     #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
     fn initialize_builtins_with_disordered_builtins() {
         let program = program![BuiltinName::range_check, BuiltinName::output];
-        let cairo_runner = cairo_runner!(program, "plain");
+        let cairo_runner = cairo_runner!(program, LayoutName::plain);
         let mut vm = vm!();
         assert!(cairo_runner.initialize_builtins(&mut vm, false).is_err());
     }
@@ -1673,7 +1667,7 @@ mod tests {
     #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
     fn initialize_builtins_missing_builtins_no_allow_missing() {
         let program = program![BuiltinName::output, BuiltinName::ecdsa];
-        let cairo_runner = cairo_runner!(program, "plain");
+        let cairo_runner = cairo_runner!(program, LayoutName::plain);
         let mut vm = vm!();
         assert_matches!(
             cairo_runner.initialize_builtins(&mut vm, false),
@@ -3822,7 +3816,7 @@ mod tests {
         .unwrap();
 
         let mut hint_processor = BuiltinHintProcessor::new_empty();
-        let mut cairo_runner = cairo_runner!(program, "all_cairo", true);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::all_cairo, true);
         let mut vm = vm!(true);
 
         let end = cairo_runner.initialize(&mut vm, false).unwrap();
@@ -3889,7 +3883,7 @@ mod tests {
             entrypoint: "main",
             trace_enabled: true,
             relocate_mem: false,
-            layout: "all_cairo",
+            layout: LayoutName::all_cairo,
             proof_mode: false,
             secure_run: Some(false),
             ..Default::default()
@@ -3907,7 +3901,7 @@ mod tests {
             entrypoint: "main",
             trace_enabled: false,
             relocate_mem: false,
-            layout: "all_cairo",
+            layout: LayoutName::all_cairo,
             proof_mode: false,
             secure_run: Some(false),
             ..Default::default()
@@ -4023,7 +4017,7 @@ mod tests {
     #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
     fn finalize_segments_run_ended_emptyproof_mode() {
         let program = program!();
-        let mut cairo_runner = cairo_runner!(program, "plain", true);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::plain, true);
         cairo_runner.program_base = Some(Relocatable::from((0, 0)));
         cairo_runner.execution_base = Some(Relocatable::from((1, 0)));
         cairo_runner.run_ended = true;
@@ -4040,7 +4034,7 @@ mod tests {
         Arc::get_mut(&mut program.shared_program_data).unwrap().data =
             vec_data![(1), (2), (3), (4), (5), (6), (7), (8)];
         //Program data len = 8
-        let mut cairo_runner = cairo_runner!(program, "plain", true);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::plain, true);
         cairo_runner.program_base = Some(Relocatable::from((0, 0)));
         cairo_runner.execution_base = Some(Relocatable::from((1, 0)));
         cairo_runner.run_ended = true;
@@ -4074,7 +4068,7 @@ mod tests {
         Arc::get_mut(&mut program.shared_program_data).unwrap().data =
             vec_data![(1), (2), (3), (4)];
         //Program data len = 4
-        let mut cairo_runner = cairo_runner!(program, "plain", true);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::plain, true);
         cairo_runner.program_base = Some(Relocatable::from((0, 0)));
         cairo_runner.execution_base = Some(Relocatable::from((1, 1)));
         cairo_runner.execution_public_memory = Some(vec![1_usize, 3_usize, 5_usize, 4_usize]);
@@ -4179,7 +4173,7 @@ mod tests {
     fn check_range_check_usage_without_builtins() {
         let program = program!();
 
-        let cairo_runner = cairo_runner!(program, "plain");
+        let cairo_runner = cairo_runner!(program, LayoutName::plain);
         let mut vm = vm!();
         vm.builtin_runners = vec![];
         vm.current_step = 10000;
@@ -4501,12 +4495,12 @@ mod tests {
     fn initialize_segments_incorrect_layout_plain_one_builtin() {
         let program = program![BuiltinName::output];
         let mut vm = vm!();
-        let cairo_runner = cairo_runner!(program, "plain");
+        let cairo_runner = cairo_runner!(program, LayoutName::plain);
         assert_eq!(
             cairo_runner.initialize_builtins(&mut vm, false),
             Err(RunnerError::NoBuiltinForInstance(Box::new((
                 HashSet::from([BuiltinName::output]),
-                String::from("plain")
+                LayoutName::plain
             ))))
         );
     }
@@ -4516,12 +4510,12 @@ mod tests {
     fn initialize_segments_incorrect_layout_plain_two_builtins() {
         let program = program![BuiltinName::output, BuiltinName::pedersen];
         let mut vm = vm!();
-        let cairo_runner = cairo_runner!(program, "plain");
+        let cairo_runner = cairo_runner!(program, LayoutName::plain);
         assert_eq!(
             cairo_runner.initialize_builtins(&mut vm, false),
             Err(RunnerError::NoBuiltinForInstance(Box::new((
                 HashSet::from([BuiltinName::output, BuiltinName::pedersen]),
-                String::from("plain")
+                LayoutName::plain
             ))))
         );
     }
@@ -4531,12 +4525,12 @@ mod tests {
     fn initialize_segments_incorrect_layout_small_two_builtins() {
         let program = program![BuiltinName::output, BuiltinName::bitwise];
         let mut vm = vm!();
-        let cairo_runner = cairo_runner!(program, "small");
+        let cairo_runner = cairo_runner!(program, LayoutName::small);
         assert_eq!(
             cairo_runner.initialize_builtins(&mut vm, false),
             Err(RunnerError::NoBuiltinForInstance(Box::new((
                 HashSet::from([BuiltinName::bitwise]),
-                String::from("small")
+                LayoutName::small,
             ))))
         );
     }
@@ -4700,7 +4694,7 @@ mod tests {
         Arc::get_mut(&mut program.shared_program_data).unwrap().data =
             vec_data![(1), (2), (3), (4), (5), (6), (7), (8)];
         //Program data len = 8
-        let mut cairo_runner = cairo_runner!(program, "plain", true);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::plain, true);
         cairo_runner.program_base = Some(Relocatable::from((0, 0)));
         cairo_runner.execution_base = Some(Relocatable::from((1, 0)));
         cairo_runner.run_ended = true;
@@ -4724,7 +4718,7 @@ mod tests {
         Arc::get_mut(&mut program.shared_program_data).unwrap().data =
             vec_data![(1), (2), (3), (4), (5), (6), (7), (8)];
         //Program data len = 8
-        let mut cairo_runner = cairo_runner!(program, "plain", true);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::plain, true);
         cairo_runner.program_base = Some(Relocatable::from((0, 0)));
         cairo_runner.execution_base = Some(Relocatable::from((1, 0)));
         cairo_runner.run_ended = false;
@@ -4742,7 +4736,7 @@ mod tests {
         Arc::get_mut(&mut program.shared_program_data).unwrap().data =
             vec_data![(1), (2), (3), (4), (5), (6), (7), (8)];
         //Program data len = 8
-        let mut cairo_runner = cairo_runner!(program, "plain", true);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::plain, true);
         cairo_runner.program_base = Some(Relocatable::from((0, 0)));
         cairo_runner.execution_base = Some(Relocatable::from((1, 0)));
         cairo_runner.run_ended = true;
@@ -4761,7 +4755,7 @@ mod tests {
         Arc::get_mut(&mut program.shared_program_data).unwrap().data =
             vec_data![(1), (2), (3), (4), (5), (6), (7), (8)];
         //Program data len = 8
-        let mut cairo_runner = cairo_runner!(program, "all_cairo", true);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::all_cairo, true);
         cairo_runner.program_base = Some(Relocatable::from((0, 0)));
         cairo_runner.execution_base = Some(Relocatable::from((1, 0)));
         cairo_runner.run_ended = true;
@@ -4792,7 +4786,7 @@ mod tests {
         Arc::get_mut(&mut program.shared_program_data).unwrap().data =
             vec_data![(1), (2), (3), (4), (5), (6), (7), (8)];
         //Program data len = 8
-        let mut cairo_runner = cairo_runner!(program, "all_cairo", true);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::all_cairo, true);
         cairo_runner.program_base = Some(Relocatable::from((0, 0)));
         cairo_runner.execution_base = Some(Relocatable::from((1, 0)));
         cairo_runner.run_ended = true;
@@ -4823,7 +4817,7 @@ mod tests {
         Arc::get_mut(&mut program.shared_program_data).unwrap().data =
             vec_data![(1), (2), (3), (4), (5), (6), (7), (8)];
         //Program data len = 8
-        let mut cairo_runner = cairo_runner!(program, "all_cairo", true);
+        let mut cairo_runner = cairo_runner!(program, LayoutName::all_cairo, true);
         cairo_runner.program_base = Some(Relocatable::from((0, 0)));
         cairo_runner.execution_base = Some(Relocatable::from((1, 0)));
         cairo_runner.run_ended = true;
@@ -5334,14 +5328,15 @@ mod tests {
 
     #[test]
     fn get_cairo_pie_no_program_base() {
-        let runner = CairoRunner::new(&Program::default(), "all_cairo", false).unwrap();
+        let runner = CairoRunner::new(&Program::default(), LayoutName::all_cairo, false).unwrap();
         let vm = vm!();
         assert_eq!(runner.get_cairo_pie(&vm), Err(RunnerError::NoProgBase))
     }
 
     #[test]
     fn get_cairo_pie_no_execution_base() {
-        let mut runner = CairoRunner::new(&Program::default(), "all_cairo", false).unwrap();
+        let mut runner =
+            CairoRunner::new(&Program::default(), LayoutName::all_cairo, false).unwrap();
         let vm = vm!();
         runner.program_base = Some(Relocatable::from((0, 0)));
         assert_eq!(runner.get_cairo_pie(&vm), Err(RunnerError::NoExecBase))
@@ -5349,7 +5344,8 @@ mod tests {
 
     #[test]
     fn get_cairo_pie_no_segment_sizes() {
-        let mut runner = CairoRunner::new(&Program::default(), "all_cairo", false).unwrap();
+        let mut runner =
+            CairoRunner::new(&Program::default(), LayoutName::all_cairo, false).unwrap();
         let mut vm = vm!();
         runner.program_base = Some(Relocatable::from((0, 0)));
         runner.execution_base = Some(Relocatable::from((1, 0)));
@@ -5369,7 +5365,8 @@ mod tests {
 
     #[test]
     fn get_cairo_pie_ret_pc_segment_size_not_zero() {
-        let mut runner = CairoRunner::new(&Program::default(), "all_cairo", false).unwrap();
+        let mut runner =
+            CairoRunner::new(&Program::default(), LayoutName::all_cairo, false).unwrap();
         let mut vm = vm!();
         runner.program_base = Some(Relocatable::from((0, 0)));
         runner.execution_base = Some(Relocatable::from((1, 0)));
@@ -5391,7 +5388,8 @@ mod tests {
 
     #[test]
     fn get_cairo_pie_program_base_offset_not_zero() {
-        let mut runner = CairoRunner::new(&Program::default(), "all_cairo", false).unwrap();
+        let mut runner =
+            CairoRunner::new(&Program::default(), LayoutName::all_cairo, false).unwrap();
         let mut vm = vm!();
         runner.program_base = Some(Relocatable::from((0, 1)));
         runner.execution_base = Some(Relocatable::from((1, 0)));
@@ -5413,7 +5411,8 @@ mod tests {
 
     #[test]
     fn get_cairo_pie_execution_base_offset_not_zero() {
-        let mut runner = CairoRunner::new(&Program::default(), "all_cairo", false).unwrap();
+        let mut runner =
+            CairoRunner::new(&Program::default(), LayoutName::all_cairo, false).unwrap();
         let mut vm = vm!();
         runner.program_base = Some(Relocatable::from((0, 0)));
         runner.execution_base = Some(Relocatable::from((1, 1)));
@@ -5435,7 +5434,8 @@ mod tests {
 
     #[test]
     fn get_cairo_pie_ret_fp_offset_not_zero() {
-        let mut runner = CairoRunner::new(&Program::default(), "all_cairo", false).unwrap();
+        let mut runner =
+            CairoRunner::new(&Program::default(), LayoutName::all_cairo, false).unwrap();
         let mut vm = vm!();
         runner.program_base = Some(Relocatable::from((0, 0)));
         runner.execution_base = Some(Relocatable::from((1, 0)));
@@ -5457,7 +5457,8 @@ mod tests {
 
     #[test]
     fn get_cairo_pie_ret_pc_offset_not_zero() {
-        let mut runner = CairoRunner::new(&Program::default(), "all_cairo", false).unwrap();
+        let mut runner =
+            CairoRunner::new(&Program::default(), LayoutName::all_cairo, false).unwrap();
         let mut vm = vm!();
         runner.program_base = Some(Relocatable::from((0, 0)));
         runner.execution_base = Some(Relocatable::from((1, 0)));
@@ -5479,7 +5480,8 @@ mod tests {
 
     #[test]
     fn get_cairo_pie_ok() {
-        let mut runner = CairoRunner::new(&Program::default(), "all_cairo", false).unwrap();
+        let mut runner =
+            CairoRunner::new(&Program::default(), LayoutName::all_cairo, false).unwrap();
         let mut vm = vm!();
         runner.program_base = Some(Relocatable::from((0, 0)));
         runner.execution_base = Some(Relocatable::from((1, 0)));
@@ -5503,7 +5505,7 @@ mod tests {
             program_content,
             &CairoRunConfig {
                 proof_mode: true,
-                layout: "all_cairo",
+                layout: LayoutName::all_cairo,
                 ..Default::default()
             },
             &mut BuiltinHintProcessor::new_empty(),